+++ rectools/metrics/popularity.py
@@ -16,6 +16,7 @@
 
 import typing as tp
 
+import attr
 import pandas as pd
 
 from rectools import Columns
@@ -23,6 +24,7 @@ from rectools.metrics.base import MetricAtK
 from rectools.utils import select_by_type
 
 
+@attr.s
 class AvgRecPopularity(MetricAtK):
     r"""
     Average Recommendations Popularity metric.
@@ -32,17 +34,24 @@ class AvgRecPopularity(MetricAtK):
     with this item.
 
     .. math::
-        ARP@k = \frac{1}{\left|U_{t}\right|}\sum_{u\in U_{t}^{}}\frac{\sum_{i\in L_{u}}\phi (i)}{\left| L_{u} \right |}
+        ARP@k = \frac{1}{|U_{t}|}\sum_{u\in U_{t}^{}}\frac{\sum_{i\in L_{u}}\phi (i)}{|L_{u}|}
+    .. math::
+        Normalized ARP@k = \frac{1}{|U_t|}\sum_{u\in U_t^{}}\frac{(\sum_{i\in L_u}\phi(i))/|interactions|}{|L_u|}
 
     where
-    :math:`\phi (i)` is the number of previous interactions with item i.
-    :math:`|U_{t}|` is the number of users in the test set.
-    :math:`L_{u}` is the list of top k recommended items for user u.
+        - :math:`\phi (i)` is the number of previous interactions with item i;
+        - :math:`|U_{t}|` is the number of users in the test set;
+        - :math:`|interactions|` is the total number of interactions;
+        - :math:`L_{u}` is the list of top k recommended items for user u.
 
     Parameters
     ----------
     k : int
         Number of items at the top of recommendations list that will be used to calculate metric.
+    normalize: bool
+        Flag, which says whether to normalize metric or not.
+        Normalization is done on total items popularity. This gives a probabilistic
+        interpretation of the metric that can be easily applied to any data.
 
     Examples
     --------
@@ -63,7 +72,11 @@ class AvgRecPopularity(MetricAtK):
     array([3., 1., 1.])
     >>> AvgRecPopularity(k=3).calc_per_user(reco, prev_interactions).values
     array([2.5, 2. , 1.5])
+    >>> AvgRecPopularity(k=3, normalize=True).calc_per_user(reco, prev_interactions).values
+    array([0.41666667, 0.33333333, 0.25        ])
     """
+
+    normalize: bool = attr.ib(default=False)
 
     def calc(self, reco: pd.DataFrame, prev_interactions: pd.DataFrame) -> float:
         """
@@ -106,7 +119,7 @@ class AvgRecPopularity(MetricAtK):
         pd.Series
             Values of metric (index - user id, values - metric value for every user).
         """
-        item_popularity = prev_interactions[Columns.Item].value_counts()
+        item_popularity = prev_interactions[Columns.Item].value_counts(normalize=self.normalize)
         item_popularity.name = "popularity"
 
         reco_k = reco.query(f"{Columns.Rank} <= @self.k")